chore(fdc): Refactor data connect error logic
